LayoutTests/fast/css/rule-selector-overflow-expected.txt

This test tests and documents the behavior of CSS style rules with a massive number of selectors. Rules with >8192 selectors get split into multiple rules at the parsing stage. Setting a rule's selectorText via CSSOM will do nothing if there are more than 8192 selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rule().selectorText = selectorListWithLength(1); rule().selectorText is selectorListWithLength(1)
PASS rule().selectorText = selectorListWithLength(8192); rule().selectorText is selectorListWithLength(8192)
PASS rule().selectorText = '.reset'; rule().selectorText is '.reset'
PASS rule().selectorText = selectorListWithLength(8193); rule().selectorText is '.reset'
PASS rule().selectorText = '.reset'; rule().selectorText is '.reset'
PASS rule().selectorText = selectorListWithLength(8193); sheet().rules.length is 1
PASS rule().selectorText = selectorListWithLength(8192); rule().selectorText is selectorListWithLength(8192)
PASS rule().selectorText = selectorListWithLength(8192); sheet().rules.length is 1
PASS rule().selectorText = '.reset'; rule().selectorText is '.reset'
PASS rule().selectorText = selectorListWithLength(8193); rule().selectorText is '.reset'
PASS rule().selectorText = selectorListWithLength(8193); sheet().rules.length is 1
PASS styleElement.innerText = styleSheetWithSelectorLength(1); rule().selectorText is selectorListWithLength(1)
PASS styleElement.innerText = styleSheetWithSelectorLength(8192); rule().selectorText is selectorListWithLength(8192)
PASS styleElement.innerText = styleSheetWithSelectorLength(8192); sheet().rules.length is 1
PASS styleElement.innerText = styleSheetWithSelectorLength(8193); rule().selectorText is selectorListWithLength(8192)
PASS styleElement.innerText = styleSheetWithSelectorLength(8193); sheet().rules.length is 2
PASS styleElement.innerText = styleSheetWithSelectorLength(16384); sheet().rules.length is 2
PASS styleElement.innerText = styleSheetWithSelectorLength(16385); sheet().rules.length is 3
PASS successfullyParsed is true

TEST COMPLETE
